pn_requirements.txt
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-JWT==0.3.2
Flask-JWT-Extended==3.23.0
Flask-SQLAlchemy==2.4.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mongoengine==0.18.2
Naked==0.1.31
psycopg2==2.8.3
pycrypto==2.6.1
PyJWT==1.4.2
pymongo==3.9.0
pypng==0.0.20
PyQRCode==1.2.1
PyYAML==5.1.2
requests==2.22.0
shellescape==3.4.1
six==1.12.0
SQLAlchemy==1.3.8
urllib3==1.25.3
Werkzeug==0.15.6
